wip contact page refactor
