+++ core/app/[locale]/(default)/contentful/contact/_components/contact-form.tsx
@@ -8,11 +8,10 @@ import { Input } from '@/vibes/soul/form/input';
 import { Label } from '@/vibes/soul/form/label';
 import { Select } from '@/vibes/soul/form/select';
 import { Button } from '@/vibes/soul/primitives/button';
-import { Modal } from '@/vibes/soul/primitives/modal';
 import { toast } from '@/vibes/soul/primitives/toaster';
 
 import { submitForm } from '../_actions/submit-form';
-import { TicketField } from '../page';
+import { TicketField } from '../form/page';
 
 export interface FormState {
   errors: Record<string, string[]> | null;
@@ -48,7 +47,6 @@ export const ContactForm = ({ fields }: { fields: TicketField[] }) => {
     handleSubmitWithParam,
     initialState,
   );
-  const [isOpen, setIsOpen] = useState(false);
   const [formFields, setFormFields] = useState<TicketField[]>(fields);
 
   const handleSelectChange = (field: TicketField, value: string) => {
@@ -81,7 +79,6 @@ export const ContactForm = ({ fields }: { fields: TicketField[] }) => {
 
   useEffect(() => {
     if (formState.success) {
-      setIsOpen(false);
       toast.success(
         <div
           dangerouslySetInnerHTML={{
@@ -93,101 +90,73 @@ export const ContactForm = ({ fields }: { fields: TicketField[] }) => {
     }
   }, [formState.success]);
 
-  useEffect(() => {
-    if (formState.errors && formRef.current) {
-      formRef.current.scrollIntoView({ behavior: 'smooth' });
-    }
-  }, [formState.errors]);
-
   return (
-    <Modal
-      isOpen={isOpen}
-      setOpen={setIsOpen}
-      title="Send an email"
-      trigger={
-        <Button className="w-full md:w-auto" id="contact-form-button" size="medium">
-          Send an email
-        </Button>
-      }
+    <form
+      action={formAction}
+      className="flex max-w-lg flex-col gap-6 md:gap-4"
+      id="contact-form"
+      noValidate
+      ref={formRef}
     >
-      <form
-        action={formAction}
-        className="flex max-w-lg flex-col gap-6 md:gap-4"
-        noValidate
-        ref={formRef}
-      >
-        {formState.errors && (
-          <FieldError className="mt-4 mb-4" key="form-errors-general">
-            Form has errors, see below.
-          </FieldError>
-        )}
-        <div className="flex flex-col gap-1">
-          <Label className="text-foreground text-sm font-medium" htmlFor="email">
-            Email Address
-          </Label>
-          <Input
-            defaultValue={getFormValue(formState.formData, 'email')}
-            errors={getErrorsOrUndefined(formState.errors, 'email')}
-            id="email"
-            name="email"
-            required
-            type="email"
-          />
-        </div>
-        {formFields.map(
-          (field) =>
-            !field.hidden && (
-              <div className="flex flex-col gap-1" key={field.id}>
-                <Label
-                  className="text-foreground text-sm font-medium"
-                  htmlFor={field.id.toString()}
-                >
-                  {field.title_in_portal}
-                </Label>
-                {field.description ? (
-                  <p className="text-contrast-400 text-xs">{field.description}</p>
-                ) : null}
-                {field.custom_field_options ? (
-                  <Select
-                    defaultValue={getFormValue(formState.formData, field.id.toString())}
-                    errors={getErrorsOrUndefined(formState.errors, field.id.toString())}
-                    id={field.id.toString()}
-                    name={field.id.toString()}
-                    onValueChange={(value) => handleSelectChange(field, value)}
-                    options={field.custom_field_options.map((option) => ({
-                      label: option.name,
-                      value: option.value,
-                    }))}
-                    required={field.required}
-                  />
-                ) : (
-                  <Input
-                    defaultValue={getFormValue(formState.formData, field.id.toString())}
-                    errors={getErrorsOrUndefined(formState.errors, field.id.toString())}
-                    id={field.id.toString()}
-                    name={field.id.toString()}
-                    required={field.required}
-                    type="text"
-                  />
-                )}
-              </div>
-            ),
-        )}
-        <div className="flex justify-end gap-2 md:mt-2">
-          <Button
-            className="flex-1 md:flex-none"
-            disabled={isPending}
-            onClick={() => setIsOpen(false)}
-            size="medium"
-            variant="tertiary"
-          >
-            Cancel
-          </Button>
-          <Button className="flex-1 md:flex-none" disabled={isPending} size="medium" type="submit">
-            {isPending ? <Loader2 className="mr-2 animate-spin" size={16} /> : 'Submit'}
-          </Button>
-        </div>
-      </form>
-    </Modal>
+      {formState.errors && (
+        <FieldError className="mt-4 mb-4" key="form-errors-general">
+          Form has errors, see below.
+        </FieldError>
+      )}
+      <div className="flex flex-col gap-1">
+        <Label className="text-foreground text-sm font-medium" htmlFor="email">
+          Email Address
+        </Label>
+        <Input
+          defaultValue={getFormValue(formState.formData, 'email')}
+          errors={getErrorsOrUndefined(formState.errors, 'email')}
+          id="email"
+          name="email"
+          required
+          type="email"
+        />
+      </div>
+      {formFields.map(
+        (field) =>
+          !field.hidden && (
+            <div className="flex flex-col gap-1" key={field.id}>
+              <Label className="text-foreground text-sm font-medium" htmlFor={field.id.toString()}>
+                {field.title_in_portal}
+              </Label>
+              {field.description ? (
+                <p className="text-contrast-400 text-xs">{field.description}</p>
+              ) : null}
+              {field.custom_field_options ? (
+                <Select
+                  defaultValue={getFormValue(formState.formData, field.id.toString())}
+                  errors={getErrorsOrUndefined(formState.errors, field.id.toString())}
+                  id={field.id.toString()}
+                  name={field.id.toString()}
+                  onValueChange={(value) => handleSelectChange(field, value)}
+                  options={field.custom_field_options.map((option) => ({
+                    label: option.name,
+                    value: option.value,
+                  }))}
+                  required={field.required}
+                />
+              ) : (
+                <Input
+                  defaultValue={getFormValue(formState.formData, field.id.toString())}
+                  errors={getErrorsOrUndefined(formState.errors, field.id.toString())}
+                  id={field.id.toString()}
+                  name={field.id.toString()}
+                  required={field.required}
+                  type="text"
+                />
+              )}
+            </div>
+          ),
+      )}
+      <div className="flex justify-end gap-2 md:mt-2">
+        <Button className="flex-1 md:flex-none" disabled={isPending} size="medium" type="submit">
+          {isPending ? <Loader2 className="mr-2 animate-spin" size={16} /> : 'Submit'}
+        </Button>
+      </div>
+    </form>
   );
 };
